refactor: 불필요한 상태 제거 및 상수,함수 분리
