+++ src/components/organisms/AddScheduleForm/AddScheduleForm.module.scss
@@ -1,7 +1,7 @@
-@import '../../../../styles/variables/colors.scss';
-@import '../../../../styles/variables/fonts.scss';
+@import '../../../styles/variables/colors.scss';
+@import '../../../styles/variables/fonts.scss';
 
-.modalWrapper {
+.modalContainer {
   position: absolute;
   top: 90px;
   left: 85%;
@@ -12,20 +12,17 @@
   border-radius: 12px;
   box-shadow: 0px 2px 20px 0px #0000001a;
 }
-.modalContainer {
-  padding: 10px;
-}
 .inputField {
   display: flex;
   flex-direction: column;
+  gap: 10px;
   padding: 5px;
 }
-.inputField label {
-  margin-bottom: 10px;
+.inputTitle {
+  @include SemiBold_Body_16;
   color: $dark-font_5A;
 }
-
-.inputField input {
+.textInput {
   width: 100%;
   height: 48px;
   padding: 15px;
@@ -33,15 +30,10 @@
   border-radius: 6px;
   border: 1px solid $light-bg_F2;
   color: $dark-font_33;
+  &::placeholder {
+    color: $light-font_9E;
+  }
 }
-.inputField input::placeholder {
-  color: $light-font_9E;
-}
-
-.titleInput {
-  background-color: $white_FF !important;
-}
-
 .colorSelectBox {
   display: flex;
   justify-content: flex-end;
@@ -54,6 +46,10 @@
   border: 2px solid $white_FF;
   border-radius: 2px;
   cursor: pointer;
+}
+
+.activeColor {
+  border: 2px solid #733232;
 }
 
 .daySelectBox {
@@ -111,7 +107,7 @@
 }
 .dash {
   width: 12px;
-  height: 2px; /* 두께를 조절할 수 있습니다 */
+  height: 2px;
   background-color: $light-blue_B3;
   margin: 0 10px;
   vertical-align: middle;
