+++ src/pages/Main/TimeTable/TimeTable.module.scss
@@ -38,7 +38,7 @@
   color: $light-font_9E;
 }
 .addScheduleBtn {
-  padding: 12px 24px;
+  padding: 12px 20px;
   background-color: $dark-font_33;
   border: none;
   border-radius: 8px;
