+++ src/pages/Main/TimeTable/TimeTable.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import styles from './TimeTable.module.scss';
 import WeeklyTimeTable from './WeeklyTimeTable/WeeklyTimeTable';
-import AddLectureForm from '../../../components/organisms/AddALectureForm/AddLectureForm';
+import AddScheduleForm from '../../../components/organisms/AddScheduleForm/AddScheduleForm';
 
 const TimeTable = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -23,7 +23,7 @@ const TimeTable = () => {
         </button>
       </div>
       <WeeklyTimeTable />
-      {isOpen && <AddLectureForm isOpen={isOpen} onClose={handleCloseModal} />}
+      {isOpen && <AddScheduleForm onClose={handleCloseModal} />}
     </div>
   );
 };
