+++ src/pages/Main/TimeTable/WeeklyTimeTable/WeeklyTimeTable.tsx
@@ -1,13 +1,16 @@
 import { useQuery } from '@tanstack/react-query';
 import TimeTableItem from '../../../../components/molecules/TimeTableItem/TimeTableItem';
-import { IScheduleElement, TIMELIST } from '../../../../constants/schedule';
+import {
+  daysOfWeek,
+  IScheduleElement,
+  TIMELIST,
+} from '../../../../constants/schedule';
 import styles from './WeeklyTimeTable.module.scss';
 import { getSchedule } from '../../../../apis/schedule';
 
 const name = '건국대학교 3-1학기'; // 임시 지정
 
 const WeeklyTimeTable = () => {
-  const daysOfWeek = ['일', '월', '화', '수', '목', '금', '토'];
   const { data } = useQuery<IScheduleElement[], Error>({
     queryKey: ['schedule', name],
     queryFn: () => getSchedule(name),
